requirements.txt
Django==3.0.3
django-bootstrap-form==3.4
django-filter==2.4.0
flake8==3.9.2
requests==2.26.0
xmltodict==0.12.0
uWSGI>=2.0.18,<2.1
psycopg2>=2.7.5,<2.8.0
